back/requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.2.1
filelock==3.12.0
mysql-connector-python==8.0.33
Pillow==9.5.0
platformdirs==3.5.0
protobuf==3.20.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.23.0
djangorestframework-simplejwt==5.2.2